refactor: コンポーネント名の修正
